refactor: spying on useDispatch
