+++ src/components/colors-bar/tests/mockStore.js
